+++ flutter_modular/lib/src/delegates/modular_router_delegate.dart
@@ -25,6 +25,7 @@ class ModularRouterDelegate extends RouterDelegate<ModularRouter>
 
   @override
   ModularRouter get currentConfiguration => _router;
+  ModularRouter get lastPage => _pages.last.router;
 
   @override
   Widget build(BuildContext context) {